requirements.txt
watchdog
requests
python-dateutil